+++ bin/lib/App.js
@@ -1,13 +1,12 @@
 const util = require('util');
 const events = require('events');
 
-var App = function(logger, logFile) {
-
-  process.title = 'Lox2MQTT'; // required to restart process at OS level
-
-  var that = this;
+const App = function(logger, logFile) {
   this.logger = logger;
   this.logFile = logFile;
+  const that = this;
+
+  process.title = 'Lox2MQTT'; // required to restart process at OS level
 
   this.logger.info('Lox2MQTT ' + process.env.npm_package_version + ' started');
   this.logger.startLog('lox2mqtt', 'Lox2MQTT', this.logFile, 'Lox2MQTT started');
@@ -16,9 +15,11 @@ var App = function(logger, logFile) {
     that.logger.info('Lox2MQTT try to stop');
     that.exit(0, 'SIGINT');
   });
+
   process.on('SIGHUP', function() {
     that.exit(0, 'SIGHUP');
   });
+
   process.on('SIGTERM', function() {
     that.exit(0, 'SIGTERM');
   });